requirements.txt
address==0.1.1
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
distlib==0.3.4
Django==3.2.10
django-contrib-comments==2.1.0
django-cors-headers==3.10.1
django-countries==7.2.1
django-crispy-forms==1.13.0
django-filter==21.1
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
django-rest-swagger==2.2.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-generators==0.5.0
drf-yasg==1.20.0
filelock==3.4.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql==0.0.3
mysqlclient==2.1.0
numpy==1.22.0
openapi-codec==1.3.2
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
platformdirs==2.4.1
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.13.0
